requirements.txt
fastapi>=0.68.1
httpx>=0.22.0
oauthlib>=3.2.2
python-jose>=3.3.0
social-auth-core>=4.4.2
starlette>=0.19.1
